[fix] fixed icon display issues and bootstrap background-color override
