+++ tedx-website/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ComingSoon from "./components/ComingSoon"
 import './App.css'
-import Footer from "./footer";
+import Footer from "./components/footer";
 
 function App() {
     return (
