+++ tedx-website/src/components/ComingSoon.css
@@ -4,21 +4,8 @@ html {
     background-size: cover;
 }
 
-/* fade in animation
 body {
-    animation: fadeInAnimation ease 3s;
-    animation-iteration-count: 1;
-    animation-fill-mode: forwards;
-} */
-
-@keyframes fadeInAnimation {
-    0% {
-        opacity: 0;
-    }
-
-    100% {
-        opacity: 1;
-    }
+    background-color: transparent !important;
 }
 
 .content-section {
