+++ tedx-website/src/components/footer.css
@@ -1,0 +1,25 @@
+#footer {
+  position: fixed;
+  bottom: 0;
+  right: 0;
+  left: 0;
+  display: flex;
+  justify-content: space-between;
+  font-size: 12pt;
+}
+
+/* <>Margin */
+#leftContent {
+  color: #FBFBFB;
+  margin-left: 15px;
+}
+
+.icons {
+  padding-left: 10px;
+  text-decoration: none;
+}
+
+/* <>Margin */
+#rightContent {
+  margin-right: 16px;
+}
